pip-requirements.txt
jenkins-job-builder==1.6.1
multi-key-dict==2.0.3
pbr==1.10.0
python-jenkins==0.4.13
PyYAML==3.12
simplejson==3.10.0
six==1.10.0
